refactor: move code to common package
