+++ packages/espressocash_app/lib/features/activities/widgets/odp_tile.dart
@@ -1,12 +1,11 @@
 import 'package:dfunc/dfunc.dart';
+import 'package:espressocash_common/espressocash_common.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../gen/assets.gen.dart';
-import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
 import '../../../routing.dart';
 import '../../../utils/extensions.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../outgoing_direct_payments/screens/odp_details_screen.dart';
 import '../models/activity.dart';
 import 'activity_tile.dart';
@@ -21,8 +20,7 @@ class ODPTile extends StatelessWidget {
         title: context.l10n.sentDirectly,
         timestamp: context.formatDate(activity.created),
         icon: Assets.icons.paymentIcon.svg(),
-        outgoingAmount:
-            activity.data.amount.format(DeviceLocale.localeOf(context)),
+        outgoingAmount: activity.data.amount.format(context.locale),
         status: activity.data.status.map(
           txCreated: always(CpActivityTileStatus.inProgress),
           txSent: always(CpActivityTileStatus.inProgress),
