+++ packages/espressocash_app/lib/features/activities/widgets/outgoing_dln_tile.dart
@@ -1,12 +1,11 @@
 import 'package:dfunc/dfunc.dart';
+import 'package:espressocash_common/espressocash_common.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../gen/assets.gen.dart';
-import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
 import '../../../routing.dart';
 import '../../../utils/extensions.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../outgoing_dln_payments/screens/details_screen.dart';
 import '../models/activity.dart';
 import 'activity_tile.dart';
@@ -22,7 +21,7 @@ class OutgoingDlnTile extends StatelessWidget {
         timestamp: context.formatDate(activity.created),
         icon: Assets.icons.paymentIcon.svg(),
         outgoingAmount: activity.data.payment.inputAmount
-            .format(DeviceLocale.localeOf(context), maxDecimals: 2),
+            .format(context.locale, maxDecimals: 2),
         status: activity.data.status.map(
           txCreated: always(CpActivityTileStatus.inProgress),
           txSent: always(CpActivityTileStatus.inProgress),
