+++ packages/espressocash_app/lib/features/conversion_rates/widgets/amount_with_equivalent.dart
@@ -3,7 +3,6 @@ import 'package:espressocash_common/espressocash_common.dart';
 import 'package:flutter/material.dart';
 
 import '../../../l10n/decimal_separator.dart';
-import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../ui/chip.dart';
 import '../../../ui/colors.dart';
@@ -12,7 +11,6 @@ import '../../../ui/shake.dart';
 import '../../../ui/usdc_info.dart';
 import '../../wallet_flow/widgets/extensions.dart';
 import '../services/amount_ext.dart';
-import 'extensions.dart';
 
 class AmountWithEquivalent extends StatelessWidget {
   const AmountWithEquivalent({
