+++ packages/espressocash_app/lib/features/outgoing_direct_payments/widgets/link_listener.dart
@@ -6,12 +6,10 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:solana/solana_pay.dart';
 
-import '../../../l10n/device_locale.dart';
 import '../../../routing.dart';
 import '../../../utils/solana_pay.dart';
 import '../../conversion_rates/data/repository.dart';
 import '../../conversion_rates/services/amount_ext.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../dynamic_links/widgets/dynamic_link_handler.dart';
 import '../screens/odp_confirmation_screen.dart';
 import '../screens/odp_details_screen.dart';
