+++ packages/espressocash_app/lib/features/outgoing_dln_payments/screens/details_screen.dart
@@ -5,7 +5,6 @@ import 'package:go_router/go_router.dart';
 import 'package:recase/recase.dart';
 
 import '../../../di.dart';
-import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../ui/button.dart';
 import '../../../ui/content_padding.dart';
@@ -18,7 +17,6 @@ import '../../../ui/timeline.dart';
 import '../../../ui/web_view_screen.dart';
 import '../../../utils/extensions.dart';
 import '../../authenticated/authenticated_navigator_key.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../intercom/services/intercom_service.dart';
 import '../../transactions/widgets/transfer_progress.dart';
 import '../data/repository.dart';
