+++ packages/espressocash_app/lib/features/payment_request/widgets/formatted_amount.dart
@@ -1,6 +1,6 @@
+import 'package:espressocash_common/espressocash_common.dart';
 import 'package:flutter/widgets.dart';
 
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../tokens/token_list.dart';
 import '../models/payment_request.dart';
