+++ packages/espressocash_app/lib/features/ramp/widgets/off_ramp_confirmation.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 
 import '../../../di.dart';
 import '../../../gen/assets.gen.dart';
-import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../ui/app_bar.dart';
 import '../../../ui/back_button.dart';
@@ -13,7 +12,6 @@ import '../../../ui/colors.dart';
 import '../../../ui/content_padding.dart';
 import '../../../ui/info_widget.dart';
 import '../../../ui/theme.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../fees/models/fee_type.dart';
 import '../../fees/widgets/fee_label.dart';
 import '../services/off_ramp_order_service.dart';
