+++ packages/espressocash_app/lib/features/ramp/widgets/on_ramp_deposit_widget.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import '../../../di.dart';
-import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../ui/back_button.dart';
 import '../../../ui/button.dart';
@@ -12,7 +11,6 @@ import '../../../ui/content_padding.dart';
 import '../../../ui/rounded_rectangle.dart';
 import '../../../ui/snackbar.dart';
 import '../../../ui/theme.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../data/on_ramp_order_service.dart';
 import 'countdown_timer.dart';
