+++ packages/espressocash_app/lib/features/wallet_flow/screens/pay_screen.dart
@@ -5,12 +5,10 @@ import 'package:go_router/go_router.dart';
 import 'package:solana/solana.dart';
 
 import '../../../gen/assets.gen.dart';
-import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../routing.dart';
 import '../../authenticated/authenticated_navigator_key.dart';
 import '../../blockchain/models/blockchain.dart';
-import '../../conversion_rates/widgets/extensions.dart';
 import '../../outgoing_direct_payments/screens/odp_confirmation_screen.dart';
 import '../../outgoing_direct_payments/screens/odp_details_screen.dart';
 import '../../outgoing_direct_payments/screens/odp_input_screen.dart';
