+++ packages/espressocash_common/lib/espressocash_common.dart
@@ -1,4 +1,6 @@
 export 'src/amount.dart';
 export 'src/currency.dart';
+export 'src/device_locale.dart';
 export 'src/flow.dart';
 export 'src/token.dart';
+export 'src/utils/format_amount.dart';
