+++ packages/espressocash_common/pubspec.yaml
@@ -1,6 +1,6 @@
 name: espressocash_common
 description: Espresso Cash Common Library
-version: 1.0.0
+version: 1.0.1
 publish_to: "none"
 
 environment:
@@ -9,7 +9,10 @@ environment:
 dependencies:
   decimal: ^2.3.3
   dfunc: ^0.9.0
+  flutter:
+    sdk: flutter
   freezed_annotation: ^2.4.1
+  intl: ^0.18.1
   json_annotation: ^4.8.1
   meta: ^1.10.0
   solana:
